# admin/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ajin-blog 本机日报管理 API",
    "version": "0.1.0",
    "description": "Local-only daily-report observation API. Authentication is required for all business data. No publication commands are implemented in this version."
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "登录本机后台",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              },
              "example": {
                "password": "example-not-a-real-password"
              }
            }
          },
          "required": true,
          "description": "Validated input for login"
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Session_"
                },
                "example": {
                  "data": {
                    "csrf": "example"
                  },
                  "observed_at": "example"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                },
                "example": {}
              }
            }
          },
          "default": {
            "description": "Unexpected error; retry after checking local service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          }
        },
        "description": "登录本机后台",
        "security": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Exact loopback origin of this service"
          }
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "tags": [
          "认证"
        ],
        "summary": "读取会话",
        "operationId": "session",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Session_"
                },
                "example": {
                  "data": {
                    "csrf": "example"
                  },
                  "observed_at": "example"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error; retry after checking local service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "description": "读取会话"
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "认证"
        ],
        "summary": "退出登录",
        "operationId": "logout",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Ok_"
                },
                "example": {
                  "data": {
                    "ok": false
                  },
                  "observed_at": "example"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error; retry after checking local service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "description": "退出登录",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Exact loopback origin of this service"
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Token returned by login or session"
          }
        ]
      }
    },
    "/api/system/status": {
      "get": {
        "tags": [
          "系统"
        ],
        "summary": "读取采集状态",
        "operationId": "system_status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_System_"
                },
                "example": {
                  "data": {
                    "worker_online": false,
                    "repo_ready": false,
                    "mode": "observe",
                    "host": "example",
                    "port": 1
                  },
                  "observed_at": "example"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error; retry after checking local service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "description": "读取采集状态"
      }
    },
...
    "/api/overview": {
      "get": {
        "tags": [
          "日报"
        ],
        "summary": "今日日报与未完成记录",
        "operationId": "overview",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Overview_"
                },
                "example": {
                  "data": {
                    "today": {
                      "date": "example",
                      "status": "published",
                      "stage": "example",
                      "title": "example",
                      "reason": "example",
                      "last_event": "example",
                      "observed_at": "example",
                      "schedule": {
                        "known": null,
                        "enabled": null
                      },
                      "scheduled_at": "example",
                      "run_count": 1,
                      "latest_run_id": "example",
                      "evidence": {
                        "status": null
                      },
                      "article": null,
                      "capabilities": {
                        "retry": null,
                        "backfill": null,
                        "reason": null
                      }
                    },
                    "unfinished": []
                  },
                  "observed_at": "example"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          },
          "default": {
            "description": "Unexpected error; retry after checking local service",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "example",
                  "message": "example"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "description": "今日日报与未完成记录"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_Ok_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Ok",
            "description": "Operation result"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At",
            "description": "Server observation time in UTC"
          }
        },
        "type": "object",
        "required": [
          "data",
          "observed_at"
        ],
        "title": "Envelope[Ok]",
        "description": "Envelope_Ok_ API value"
      },
...
      "Envelope_Session_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Session",
            "description": "Operation result"
          },
          "observed_at": {
            "type": "string",
            "title": "Observed At",
            "description": "Server observation time in UTC"
          }
        },
        "type": "object",
        "required": [
          "data",
          "observed_at"
        ],
        "title": "Envelope[Session]",
        "description": "Envelope_Session_ API value"
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "code as recorded by the local service"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "message as recorded by the local service"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "Error",
        "description": "Error API value"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail",
            "description": "detail as recorded by the local service"
          }
        },
        "type": "object",
        "title": "HTTPValidationError",
        "description": "HTTPValidationError API value"
      },
      "Login": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "description": "Local administrator password; never stored in browser storage"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "Login",
        "description": "Login API value"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ajin_admin_session"
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4318"
    }
  ],
...
}